refactor: add HostWindowManager to manager SwiftUI windows
